#9 - Alterando o tipo da variável valor

+++ src/main/java/com/imd/config/mudanca/banco/command/Operacao.java
@@ -1,8 +1,10 @@
 package com.imd.config.mudanca.banco.command;
 
+import java.math.BigDecimal;
+
 import com.imd.config.mudanca.banco.domain.Conta;
 
 public interface Operacao {
 
-		public double executar(Conta contaOrigem, Conta contaDestino, int valor);
+		public double executar(Conta contaOrigem, Conta contaDestino, BigDecimal valor);
 }
